Chrome: Rename utils.py -> browser_credentials_database_path.py
