+++ monkey/agent_plugins/credentials_collectors/chrome/src/browser_credentials_database_path.py
