+++ monkey/agent_plugins/credentials_collectors/chrome/src/typedef.py
@@ -3,7 +3,7 @@ from typing import Callable, Collection, TypeAlias
 from common.credentials import Credentials
 from common.types import Event
 
-from .utils import BrowserCredentialsDatabasePath
+from .browser_credentials_database_path import BrowserCredentialsDatabasePath
 
 CredentialsDatabaseSelectorCallable: TypeAlias = Callable[
     [], Collection[BrowserCredentialsDatabasePath]
